requirements.txt
PyYAML
flatten_json
requests
setuptools
